# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/public/news": {
      "get": {
        "operationId": "NewsController_list",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public/news"
        ]
      }
    },
    "/v1/public/news/{slug}": {
      "get": {
        "operationId": "NewsController_detail",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public/news"
        ]
      }
    },
    "/v1/cms/news": {
      "get": {
        "operationId": "NewsCmsController_listAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/news"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "post": {
        "operationId": "NewsCmsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNewsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cms/news"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/v1/cms/news/{id}": {
      "get": {
        "operationId": "NewsCmsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/news"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "put": {
        "operationId": "NewsCmsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNewsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/news"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "NewsCmsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/news"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/v1/cms/news/{id}/cover": {
      "post": {
        "operationId": "NewsCmsController_uploadCover",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cms/news"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/v1/cms/news/{id}/publish": {
      "patch": {
        "operationId": "NewsCmsController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/news"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/v1/public/rover/{deviceId}/latest": {
      "get": {
        "operationId": "RoverController_latestOne",
        "summary": "Posisi terkini satu rover",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public/rover"
        ]
      }
    },
    "/v1/public/rover/{deviceId}/track": {
      "get": {
        "operationId": "RoverController_track",
        "summary": "Jejak posisi rover",
        "description": "Riwayat posisi untuk gambar rute (polyline) di maps.",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public/rover"
        ]
      }
    },
...
    "/v1/public/umkm": {
      "get": {
        "operationId": "UmkmController_list",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public/umkm"
        ]
      }
    },
    "/v1/public/umkm/{slug}": {
      "get": {
        "operationId": "UmkmController_detail",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public/umkm"
        ]
      }
    },
    "/v1/cms/umkm": {
      "get": {
        "operationId": "UmkmCmsController_listAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/umkm"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "post": {
        "operationId": "UmkmCmsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUmkmDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cms/umkm"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/v1/cms/umkm/{id}": {
      "get": {
        "operationId": "UmkmCmsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/umkm"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "put": {
        "operationId": "UmkmCmsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUmkmDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/umkm"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "UmkmCmsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/umkm"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/v1/cms/umkm/{id}/cover": {
      "post": {
        "operationId": "UmkmCmsController_uploadCover",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cms/umkm"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/v1/cms/umkm/{id}/publish": {
      "patch": {
        "operationId": "UmkmCmsController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/umkm"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/v1/cms/village": {
      "put": {
        "operationId": "VillageCmsController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVillageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/village"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/v1/cms/village/gallery": {
      "post": {
        "operationId": "VillageCmsController_addGallery",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGalleryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cms/village"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/v1/cms/village/gallery/upload": {
      "post": {
        "operationId": "VillageCmsController_uploadGallery",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cms/village"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/v1/cms/village/gallery/{id}": {
      "delete": {
        "operationId": "VillageCmsController_removeGallery",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/village"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "NAWASENA Center API",
    "description": "Backend NAWASENA Center untuk PPK Ormawa RDC UAD.\n\n## Cara pakai\n- Endpoint **public/** bisa diakses tanpa login (untuk website publik).\n- Endpoint **cms/** butuh login admin. Klik tombol **Authorize** di kanan atas, masukkan token (format: `Bearer <token>` tidak perlu, cukup token-nya).\n\n## Format response\nSemua response: `{ success: true, data: {...} }` atau `{ success: false, error: {...} }`.\n\n## Autentikasi\nLogin via Supabase Auth untuk dapat access_token, lalu Authorize di sini.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Masukkan access_token dari Supabase Auth (tanpa kata Bearer)"
      }
    },
    "schemas": {
...
      "CreateNewsDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Sosialisasi Akbar NAWASENA",
            "description": "Judul berita"
          },
          "excerpt": {
            "type": "string",
            "example": "Ringkasan singkat berita.",
            "description": "Ringkasan untuk kartu/list"
          },
          "content": {
            "type": "string",
            "example": "Isi lengkap berita minimal sepuluh karakter.",
            "description": "Konten lengkap (markdown/HTML)"
          },
          "coverUrl": {
            "type": "string",
            "example": "https://.../cover.jpg",
            "description": "URL cover (isi via upload)"
          },
          "category": {
            "type": "string",
            "example": "Kegiatan",
            "description": "Kategori berita (UI only)"
          },
          "author": {
            "type": "string",
            "example": "Tim Redaksi",
            "description": "Nama penulis (UI only)"
          }
        },
        "required": [
          "title",
          "content"
        ]
      },
      "UpdateNewsDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Sosialisasi Akbar NAWASENA",
            "description": "Judul berita"
          },
          "excerpt": {
            "type": "string",
            "example": "Ringkasan singkat berita.",
            "description": "Ringkasan untuk kartu/list"
          },
          "content": {
            "type": "string",
            "example": "Isi lengkap berita minimal sepuluh karakter.",
            "description": "Konten lengkap (markdown/HTML)"
          },
          "coverUrl": {
            "type": "string",
            "example": "https://.../cover.jpg",
            "description": "URL cover (isi via upload)"
          },
          "category": {
            "type": "string",
            "example": "Kegiatan",
            "description": "Kategori berita (UI only)"
          },
          "author": {
            "type": "string",
            "example": "Tim Redaksi",
            "description": "Nama penulis (UI only)"
          }
        }
      },
...
      "ProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Keripik Original",
            "description": "Nama produk"
          },
          "price": {
            "type": "number",
            "example": 12000,
            "description": "Harga dalam rupiah"
          },
          "unit": {
            "type": "string",
            "example": "bungkus",
            "description": "Satuan: pcs, kg, bungkus"
          },
          "stock": {
            "type": "number",
            "example": 50,
            "description": "Jumlah stok"
          },
          "photoUrl": {
            "type": "string",
            "example": "https://.../produk.jpg"
          },
          "note": {
            "type": "string",
            "example": "200gr",
            "description": "Catatan produk"
          }
        },
        "required": [
          "name",
          "price"
        ]
      },
      "CreateUmkmDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Keripik Tela Bu Sari",
            "description": "Nama usaha"
          },
          "ownerName": {
            "type": "string",
            "example": "Bu Sari",
            "description": "Nama pemilik"
          },
          "description": {
            "type": "string",
            "example": "Keripik singkong renyah aneka rasa."
          },
          "category": {
            "type": "string",
            "example": "makanan",
            "description": "makanan | kerajinan | tani | lainnya"
          },
          "coverUrl": {
            "type": "string",
            "example": "https://.../cover.jpg"
          },
          "contactPhone": {
            "type": "string",
            "example": "081234567801",
            "description": "Nomor WA/telepon"
          },
          "contactAddress": {
            "type": "string",
            "example": "Padukuhan Karangtengah"
          },
          "products": {
            "description": "Daftar produk",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateUmkmDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Keripik Tela Bu Sari",
            "description": "Nama usaha"
          },
          "ownerName": {
            "type": "string",
            "example": "Bu Sari",
            "description": "Nama pemilik"
          },
          "description": {
            "type": "string",
            "example": "Keripik singkong renyah aneka rasa."
          },
          "category": {
            "type": "string",
            "example": "makanan",
            "description": "makanan | kerajinan | tani | lainnya"
          },
          "coverUrl": {
            "type": "string",
            "example": "https://.../cover.jpg"
          },
          "contactPhone": {
            "type": "string",
            "example": "081234567801",
            "description": "Nomor WA/telepon"
          },
          "contactAddress": {
            "type": "string",
            "example": "Padukuhan Karangtengah"
          },
          "products": {
            "description": "Daftar produk",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            }
          }
        }
      },
      "UpdateVillageDto": {
        "type": "object",
        "properties": {
          "vision": {
            "type": "string",
            "example": "Mewujudkan desa smart farming."
          },
          "mission": {
            "type": "string",
            "example": "Memberdayakan pemuda dan petani."
          },
          "description": {
            "type": "string",
            "example": "Kalurahan Karangtengah, Imogiri, Bantul."
          },
          "potency": {
            "type": "object",
            "example": [
              {
                "title": "Pertanian Padi",
                "desc": "Lahan produktif."
              }
            ],
            "description": "Array potensi desa"
          },
          "stats": {
            "type": "object",
            "example": {
              "population": 6800,
              "families": 1709,
              "area_ha": 350
            },
            "description": "Statistik demografi"
          }
        }
      },
      "CreateGalleryDto": {
        "type": "object",
        "properties": {
          "imageUrl": {
            "type": "string",
            "example": "https://.../foto.jpg",
            "description": "URL gambar"
          },
          "caption": {
            "type": "string",
            "example": "Sawah Karangtengah"
          },
          "orderIndex": {
            "type": "number",
            "example": 0,
            "description": "Urutan tampil"
          }
        },
        "required": [
          "imageUrl"
        ]
      },
...
    }
  }
}